requirements.txt
# Flask and extensions
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.0

# Database
pymysql==1.1.0
mysqlclient==2.2.0
sqlalchemy~=2.0
psycopg[binary]~=3.1
pgvector~=0.2

# AI and ML
openai==0.28.0
faiss-cpu==1.7.4
numpy==1.26.4
scikit-learn==1.3.2

# Utilities
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==1.10.12
PyPDF2==3.0.1

# Development
pytest==7.4.2
black==23.9.1
flake8==6.1.0
